+++ R/setup.R
@@ -1,38 +1,38 @@
-#' Setup flint
+#' Setup flir
 #'
-#' This stores the default rules and internal files in `inst/flint`. It also
+#' This stores the default rules and internal files in `inst/flir`. It also
 #' imports `sgconfig.yml` that is used by `ast-grep`. This file must live at the
 #' root of the project and cannot be renamed.
 #'
 #' @param path Path to package or project root.
 #'
-#' @return Imports files necessary for `flint` to work but doesn't return any
+#' @return Imports files necessary for `flir` to work but doesn't return any
 #' value in R.
 #' @export
 
-setup_flint <- function(path = ".") {
-  flint_dir <- file.path(path, "flint")
+setup_flir <- function(path = ".") {
+  flir_dir <- file.path(path, "flir")
 
   ### Check dir
   if (
-    fs::dir_exists(flint_dir) &&
-      length(list.files(flint_dir, recursive = TRUE)) > 0
+    fs::dir_exists(flir_dir) &&
+      length(list.files(flir_dir, recursive = TRUE)) > 0
   ) {
-    stop("Folder `flint` already exists and is not empty.")
-  } else if (!fs::dir_exists(flint_dir)) {
-    fs::dir_create(flint_dir)
+    stop("Folder `flir` already exists and is not empty.")
+  } else if (!fs::dir_exists(flir_dir)) {
+    fs::dir_create(flir_dir)
   }
 
   ### Check buildignore
   if (fs::file_exists(".Rbuildignore")) {
-    already_in <- any(grepl("flint", readLines(".Rbuildignore", warn = FALSE)))
+    already_in <- any(grepl("flir", readLines(".Rbuildignore", warn = FALSE)))
   } else {
     already_in <- FALSE
   }
   if (!already_in) {
     cat(
-      "\n\n# flint files
-^flint$\n",
+      "\n\n# flir files
+^flir$\n",
       file = ".Rbuildignore",
       append = TRUE
     )
@@ -41,54 +41,54 @@ setup_flint <- function(path = ".") {
   ### Files
   invisible(
     fs::dir_copy(
-      system.file("rules/builtin", package = "flint"),
-      fs::path(flint_dir, "rules/builtin")
+      system.file("rules/builtin", package = "flir"),
+      fs::path(flir_dir, "rules/builtin")
     )
   )
-  if (!fs::file_exists(file.path(flint_dir, "cache_file_state.rds"))) {
-    saveRDS(NULL, file.path(flint_dir, "cache_file_state.rds"))
+  if (!fs::file_exists(file.path(flir_dir, "cache_file_state.rds"))) {
+    saveRDS(NULL, file.path(flir_dir, "cache_file_state.rds"))
   }
   config_content <- paste0(
     "keep:\n",
     paste("  -", list_linters(), collapse = "\n")
   )
-  writeLines(config_content, file.path(flint_dir, "config.yml"))
+  writeLines(config_content, file.path(flir_dir, "config.yml"))
 }
 
-#' Update the `flint` setup
+#' Update the `flir` setup
 #'
 #' @description
 #'
-#' When `flint` is updated, it can ship new built-in rules or update existing
-#' ones. `update_flint()` will automatically add those new rules to the
-#' `flint/rules/builtin` folder. Custom rules stored in `flint/rules/custom`
+#' When `flir` is updated, it can ship new built-in rules or update existing
+#' ones. `update_flir()` will automatically add those new rules to the
+#' `flir/rules/builtin` folder. Custom rules stored in `flir/rules/custom`
 #' are not affected.
 #'
-#' @inheritParams setup_flint
+#' @inheritParams setup_flir
 #'
-#' @return Can add new files in the `flint/rules` folder, doesn't return anything.
+#' @return Can add new files in the `flir/rules` folder, doesn't return anything.
 #' @export
 #'
 #' @examples
 #' \dontrun{
-#'   update_flint()
+#'   update_flir()
 #' }
-update_flint <- function(path = ".") {
-  flint_dir <- file.path(path, "flint")
+update_flir <- function(path = ".") {
+  flir_dir <- file.path(path, "flir")
   built_in_rules <- list.files(
-    fs::path(flint_dir, "rules/builtin"),
+    fs::path(flir_dir, "rules/builtin"),
     pattern = "\\.yml$"
   )
   updated_built_in_rules <- system.file(
     list.files("rules/builtin", pattern = "\\.yml$"),
-    package = "flint"
+    package = "flir"
   )
   new_built_in_rules <- setdiff(updated_built_in_rules, built_in_rules)
 
   # Copy everything so that rules that already exist can also be updated.
   fs::file_copy(
     updated_built_in_rules,
-    paste0(flint_dir, "/rules/builtin/", basename(updated_built_in_rules))
+    paste0(flir_dir, "/rules/builtin/", basename(updated_built_in_rules))
   )
 
   cli::cli_alert_success("Updated existing rules.")
@@ -97,7 +97,7 @@ update_flint <- function(path = ".") {
       "Added {length(new_built_in_rules)} rule{?s}: {gsub('\\.yml$', '', basename(new_built_in_rules))}."
     )
     cli::cli_alert_info(
-      "Don't forget to add them in flint/config.yml to use them."
+      "Don't forget to add them in flir/config.yml to use them."
     )
   }
 }