+++ tests/testthat/test-config_yml.R
@@ -1,6 +1,6 @@
 test_that("config.yml is taken into account", {
   create_local_package()
-  setup_flint()
+  setup_flir()
   fs::dir_create("tests/testthat")
 
   cat("a = 1", file = "R/foo.R")
@@ -8,7 +8,7 @@ test_that("config.yml is taken into account", {
   expect_equal(nrow(lint()), 2)
 
   # # Only keep one linter, not the one about assignment symbols
-  cat("keep:\n  - class_equals", file = "flint/config.yml")
+  cat("keep:\n  - class_equals", file = "flir/config.yml")
   expect_equal(nrow(lint()), 0)
   expect_equal(nrow(lint_dir("R")), 0)
   expect_equal(nrow(lint_package()), 0)
@@ -16,7 +16,7 @@ test_that("config.yml is taken into account", {
   # commented out linter not taken into account
   cat(
     "keep:\n  - class_equals\n#  - equal_assignment",
-    file = "flint/config.yml"
+    file = "flir/config.yml"
   )
   expect_equal(nrow(lint()), 0)
   expect_equal(nrow(lint_dir("R")), 0)
@@ -25,7 +25,7 @@ test_that("config.yml is taken into account", {
   # "exclude" field works
   cat(
     "keep:\n  - class_equals\nexclude:\n  - equal_assignment",
-    file = "flint/config.yml"
+    file = "flir/config.yml"
   )
   expect_equal(nrow(lint()), 0)
   expect_equal(nrow(lint_dir("R")), 0)
@@ -34,35 +34,35 @@ test_that("config.yml is taken into account", {
 
 test_that("config.yml errors when it doesn't contain any rule", {
   create_local_package()
-  setup_flint()
+  setup_flir()
 
   # Only keep one linter, not the one about assignment symbols
-  cat("keep:", file = "flint/config.yml")
+  cat("keep:", file = "flir/config.yml")
   expect_error(lint(), "doesn't contain any rule")
 
   # commented out linter not taken into account
-  cat("keep:\n#  - equal_assignment", file = "flint/config.yml")
+  cat("keep:\n#  - equal_assignment", file = "flir/config.yml")
   expect_error(lint(), "doesn't contain any rule")
 })
 
 test_that("config.yml errors when it contains unknown rules", {
   create_local_package()
-  setup_flint()
+  setup_flir()
 
   cat(
     "keep:\n  - equal_assignment\n  - foo\n  - bar",
-    file = "flint/config.yml"
+    file = "flir/config.yml"
   )
   expect_error(lint(), "Unknown linters: foo, bar")
 })
 
 test_that("config.yml errors when it contains duplicated rules", {
   create_local_package()
-  setup_flint()
+  setup_flir()
 
   cat(
     "keep:\n  - equal_assignment\n  - class_equals\n  - equal_assignment",
-    file = "flint/config.yml"
+    file = "flir/config.yml"
   )
   expect_error(lint(), "the following linters are duplicated: equal_assignment")
 })