+++ tests/testthat/test-setup_flint_gha.R
@@ -1,8 +1,8 @@
-test_that("setup_flint_gha basic use works", {
+test_that("setup_flir_gha basic use works", {
   create_local_package()
-  expect_no_error(suppressMessages(setup_flint_gha()))
-  expect_true(file.exists(".github/workflows/flint.yaml"))
+  expect_no_error(suppressMessages(setup_flir_gha()))
+  expect_true(file.exists(".github/workflows/flir.yaml"))
 
-  expect_error(setup_flint_gha(), "file already exists")
-  expect_no_error(setup_flint_gha(overwrite = TRUE))
+  expect_error(setup_flir_gha(), "file already exists")
+  expect_no_error(setup_flir_gha(overwrite = TRUE))
 })